refactor badge status
